chore: convert files to ts
