+++ package.json
@@ -12,14 +12,14 @@
 	},
 	"packageManager": "yarn@4.6.0",
 	"scripts": {
-		"start": "node server.js",
-		"dev": "nodemon server.js",
-		"refresh": "bash ./database/database.sh",
-		"migrate": "knex migrate:latest --knexfile=./database/knexfile --verbose",
-		"seed": "knex seed:run --knexfile=./database/knexfile --verbose",
-		"rollback": "knex migrate:rollback --knexfile=./database/knexfile --verbose",
-		"make:seeder": "knex seed:make --knexfile=./database/knexfile",
-		"make:migration": "knex migrate:make --knexfile=./database/knexfile",
+		"build": "tsc -p tsconfig.json",
+		"start": "tsx src/server.ts",
+		"dev": "tsx watch src/server.ts",
+		"migrate": "knex migrate:latest --knexfile=./src/database/knexfile --verbose",
+		"seed": "knex seed:run --knexfile=./src/database/knexfile --verbose",
+		"rollback": "knex migrate:rollback --knexfile=./src/database/knexfile --verbose",
+		"make:seeder": "knex seed:make --knexfile=./src/database/knexfile",
+		"make:migration": "knex migrate:make --knexfile=./src/database/knexfile",
 		"fix": "yarn dlx @biomejs/biome check --apply .",
 		"check": "yarn dlx @biomejs/biome check .",
 		"deps": "yarn upgrade-interactive"
@@ -51,7 +51,6 @@
 		"@biomejs/biome": "1.9.4",
 		"@fastify/type-provider-json-schema-to-ts": "^5.0.0",
 		"@types/node": "^22.13.5",
-		"nodemon": "^3.1.9",
 		"pino-pretty": "^13.0.0",
 		"tsx": "^4.19.3",
 		"typescript": "^5.7.3"
