+++ src/plugins/s3object.ts
