+++ tsconfig.json
@@ -1,0 +1,28 @@
+{
+	"include": ["src/**/*.ts"],
+	"exclude": ["node_modules", "test/**/*.ts"],
+	"compilerOptions": {
+		"target": "es2024",
+		"module": "NodeNext",
+		"outDir": "dist",
+		"esModuleInterop": true,
+		"strict": true,
+		"skipLibCheck": true,
+		// "resolveJsonModule": true,
+		"removeComments": true,
+		"noUnusedLocals": true,
+		"isolatedModules": true,
+		"noFallthroughCasesInSwitch": true,
+		"forceConsistentCasingInFileNames": true,
+		"typeRoots": ["node", "@types"],
+		"baseUrl": "./src",
+		"paths": {
+			"@app/*": ["app/*"],
+			"@plugins/*": ["plugins/*"],
+			"@utility/*": ["utility/*"],
+			"@auth/*": ["app/auth/*"],
+			"@gallery/*": ["app/gallery/*"],
+			"@mail/*": ["app/mail/*"]
+		}
+	}
+}
